Refactor API for check and move cards
To make the API more flexible, I added APIs to get
card stack or stacks directly: waste(), foundation(),
tableau(), column(index).

Now move action takes 2 arguments: from stacks and to stacks.
The move API handles if provided argument is a single stack
or multiple stacks. Now player can specify from which stack/stacks
moving to which stack/stacks.

ALso introduced check_move API to check if it is valid to move
from stack/stacks to stack/stacks. The arguments are the same
as the move API. A change in behaviour is that if move API is called
without checking first, and move didn't apply, I raise an exception.

Finally, an API check_size is introduced to check the size of the
stack of cards, mostly for checking the size of the column.